format: tests

+++ test/SimpleAccount_BSC.t.sol
@@ -56,18 +56,18 @@ contract SimpleAccountBscTest is Test {
 
         console2.log("SimpleAccount wallet created at:", address(simpleAccount));
     }
-    
+
     function debugCrossChainOperation(UserOperation memory sourceUserOp) private {
         // slice post signature to retrieve the x-chain encoded Intent
         bytes memory sourceSigXChainCalldata = sourceUserOp.signature._slice(65, sourceUserOp.signature.length);
         XChainLib.xCallData memory xcd = this.parseXElems(sourceSigXChainCalldata);
-        assertEq(uint(xcd.opType), uint(XChainLib.OpType.CrossChain), "OpType should be CrossChain");
+        assertEq(uint256(xcd.opType), uint256(XChainLib.OpType.CrossChain), "OpType should be CrossChain");
         // Add more debugging logic here
     }
-    
+
     function parseXElems(bytes calldata callData) external pure returns (XChainLib.xCallData memory) {
         return XChainLib.parseXElems(callData);
-    }    
+    }
 
     function createIntent() internal pure returns (bytes memory) {
         return bytes(
